Changed float32 to decimal type for improved calculation precision

+++ CalculationLogic/Models/CaseDrainState.fs
@@ -3,27 +3,27 @@
 module CaseDrainState =
 
     type T = {
-        mutable rpm : float32
-        mutable cCRev :float32
-        mutable efficency :float32
-        mutable cCMinOut : float32
-        mutable lMinOut : float32
-        mutable area : float32
-        mutable distance : float32
-        mutable second : float32
-        mutable caseDrainGpmOut : float32
-        mutable cycleTimeGpmOut : float32
+        mutable rpm : decimal
+        mutable cCRev :decimal
+        mutable efficency :decimal
+        mutable cCMinOut : decimal
+        mutable lMinOut : decimal
+        mutable area : decimal
+        mutable distance : decimal
+        mutable second : decimal
+        mutable caseDrainGpmOut : decimal
+        mutable cycleTimeGpmOut : decimal
     }
 
     let Init = {
-        rpm = 0.0f;
-        cCRev = 0.0f;
-        efficency = 0.0f;
-        cCMinOut = 0.0f;
-        lMinOut = 0.0f;
-        area = 0.0f;
-        distance = 0.0f;
-        second = 0.0f;
-        caseDrainGpmOut = 0.0f;
-        cycleTimeGpmOut = 0.0f;
+        rpm = 0.0m;
+        cCRev = 0.0m;
+        efficency = 0.0m;
+        cCMinOut = 0.0m;
+        lMinOut = 0.0m;
+        area = 0.0m;
+        distance = 0.0m;
+        second = 0.0m;
+        caseDrainGpmOut = 0.0m;
+        cycleTimeGpmOut = 0.0m;
     }
